Домашняя работа 2.12 Тестирование

+++ src/main/java/pro/sky/java/course2/calculator/service/CalculatorService.java
@@ -9,6 +9,6 @@ public interface CalculatorService {
 
     int multiply(int num1, int num2);
 
-    double divide(double num1, double num2);
+    double divide(int num1, int num2);
 
 }
